requirements.txt
streamlit==1.42.0
sentence-transformers
faiss-cpu
pandas
numpy
requests
beautifulsoup4
lxml
fastapi
uvicorn
